+++ src/lib/Ppu/types/Tile_Types.ts
@@ -1,0 +1,4 @@
+export interface ICoordinates {
+    x: number;
+    y: number;
+}
